Move data getters to their own file
