+++ src/constants.py
@@ -43,7 +43,7 @@ def _get_filetypes():
 #################################
 # Version
 
-VERSION = "0.4.4"
+VERSION = "0.4.5"
 
 #################################
 # Environment spesific constants
